generator_backend/requirements.txt
django==5.2.2
djangorestframework==3.16.0
firebase-admin==6.9.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0 